gpu_raster_fuzzer: initialize stub ctx
The context is supposed to hold offscreen surfaces now. And given
the SharedContextState is expected to work only with offscreen
surfaces, the context it gets must hold an offscreen surface.

To fix that, the stub ctx must be initialized by the fuzzer.

This patch also cleans up the fuzzer a bit so that it doesn't
hold the surface

Cq-Include-Trybots: luci.chromium.try:linux-libfuzzer-asan-rel
Bug: 40281108, 329188915
Change-Id: I4768a5bbd18da4d1e203ec702adab44e310ed48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362057
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1272010}