[Files F2] Don't restore focus on tree items that are disabled
The OneDrive tree item gets disabled when the device goes offline. The
directory tree runs `renderRoots` as the volume list is updated, this
eventually attempts to restoreFocus however this throws a console.assert
as the item can't be focused due to being disabled. Let's check if the
item is disabled before attempting to restore focus.

Fixed: b/329328136
Change-Id: I05384aa773b9b9114c6f2c9fd9010ef3b40f9a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366424
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Wenbo Jie <wenbojie@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272011}