[Telemetry] Remove empty routine finished detail
Sync the change in mojom from CL:5364376. The [MinVersion] attributes
are modified because the changes in Mojo definition is not backward
compatible.

The corresponding struct `TelemetryDiagnosticVolumeButtonRoutineDetail`
in the crosapi is kept to meet the backward-compatibility requirement
for lacros.

Before this change, the routine finished events are emitted based on the
type of routine detail in the finished state. However, that no longer
works works after empty details are removed. Added a temporarily
workaround to support routines that have no details.

Bug: b:328729519, b:320394027
Test: browser_tests -ozone-platform=headless
Change-Id: Icdf7247d774810db690c48e7b7bb6cbb3639783f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369513
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272034}