[Quick Delete] Block all promos during the post-delete experience
This fixes a bug where the Restore tabs promo gets triggered on the tab
switcher during the post Quick Delete experience.

This handles it by having the QuickDeleteController acquire a display lock from the in-product help tracker to ensure no IpH is triggered during this time. The lock is released after the post-deletion experience is completed.

Bug: b/327252891
Change-Id: Ic387ff992567993ea330cd32c08baa25bbfb5fb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5368848
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Zaina Al-Mashni <zalmashni@google.com>
Cr-Commit-Position: refs/heads/main@{#1272066}