webview: Add metric for browsing intent fallback
If a developer _does not_ provide a WebViewClient, we will send browsing
intents if one is opened in a page.

This immediately changes if they add an empty WebViewClient, or more
likely, add a WebViewClient to potentially do something else.

We would like to explore if removing this could be an issue for anyone
so adding a metric to track this.

I tested that this showed up in chrome://histograms

Bug: 329046730
Change-Id: Ia96e7aeebe7f942d246b7cf0072af7b41185ebb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5361614
Auto-Submit: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272079}