Start post_login_deferred_task_runner_ immediately on ash restarting.
Currently, LoginUnlockThroughputRecorder has a bug on ash restartin.
So, it's partially running, but it didn't complete (i.e. no UMA is
output).
The end result is what we want, i.e., on restarting we don't want
UMA, but we have to start post_login_deferred_task_runner_,
otherwise the task scheduled there will be stuck.

This CL adds workaround, that is, on restarting, it immediate start
post_login_deferred_task_runner_.
More fix for the root cause will require some large refactoring,
and that'd be done as a part of b/328339021.

BUG=b/326065955
TEST=Tryjob. Run locally with let ash crash.

Change-Id: Id8f601d557fe445d12bfd9f889a63c1a5e8cd115
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362093
Reviewed-by: Alexandre Marciano Gimenez <raging@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272086}