iwa: Replace `IsolatedWebAppLocation` by install source mechanism
The main goal of this CL is to delete `IsolatedWebAppLocation`. Until
now, it is the main way an IWA's location / source (be it a bundle, or a
proxy) is represented and passed around in code. However, it had many
issues, such as:
- When representing a bundle, you always have to choose between
  `DevModeBundle` and `InstalledBundle`, regardless of whether or not
  your code or API cared about the difference.
- Some code may only work with bundles - until now, it needed to `CHECK`
  or have other error handling in case it is called with a proxy-based
  IWA (and vice versa).
- Similarly, some code may only work with dev mode IWAs - until now, it
  needed to have error handling in case it was called with a
  non-dev-mode location like `InstalledBundle`.
- Since `IsolatedWebAppLocation` was just an alias for a
  `absl::variant`, it was not possible to forward-declare it.
- We also tied whether or not a bundle is moved/copied/referenced to
  `InstalledBundle` and `DevModeBundle`, which lead to us not being able
  to express "copy the bundle when installed from the installer", but
  "move the bundle when installed from policy", without hacks.

As a replacement for `IsolatedWebAppLocation`, we introduce the concept
of an IWA "source". At its core, a "source" contains all information
necessary to read data from an IWA:
- `IwaSourceProxy` represents an IWA accessed through an HTTP proxy.
- `IwaSourceBundle` represents an IWA accessed from a Signed Web Bundle.
- `IwaSource` is a thin wrapper around an `absl::variant` of the former
  two sources.

However, oftentimes code needs to also know whether the IWA represented
by the source is a dev mode or non-dev-mode (= "prod mode") IWA. While
this information could be represented by an `IwaSource` in combination
with a "is-dev-mode" `bool`, this can lead to invalid states, as it
would allow an `IwaSource` holding an `IwaSourceProxy` even when the
separate `bool` was `false` (= "prod mode"). Code would then need to
either `CHECK` or somehow handle this error. Instead, we introduce more
classes that combine the aforementioned basic sources with information
about dev mode:
- `IwaSourceBundleWithMode` represents an IWA accessed from a Signed Web
  Bundle. Its mode can be queried via a `dev_mode` getter.
- `IwaSourceBundleDevMode` represents a dev mode IWA accessed from a
  Signed Web Bundle.
- `IwaSourceBundleProdMode` represents a prod mode IWA accessed from a
  Signed Web Bundle.

As `IwaSourceProxy` always represents a dev mode IWA, we do not
introduce additional classes for it.

- `IwaSourceWithMode` is a thin wrapper around an `absl::variant` of
  `IwaSourceBundleWithMode` and `IwaSourceProxy`.
- `IwaSourceDevMode` is a thin wrapper around an `absl::variant` of
  `IwaSourceBundleDevMode` and `IwaSourceProxy`.
- `IwaSourceProdMode` is a thin wrapper around an `absl::variant` of
  `IwaSourceBundleProdMode`. (`IwaSourceProxy` is always dev mode and
  thus not part of this variant.)

With these additional classes, we are now able to write type-safe code
that doesn't need to have the error handling it needed before: Code that
only deals with prod mode bundles can use `IwaSourceBundleProdMode`.
Code that only deals with dev mode IWAs can use `IwaSourceDevMode`. Code
that deals with both dev mode and prod mode IWAs can use
`IwaSourceWithMode`. Code that deals with bundles but does not care
about their mode can use `IwaSourceBundle`, etc.

This CL also decouples the input of the `InstallIsolatedWebAppCommand`
from `IsolatedWebAppLocation`. In addition to the mode of the IWA, the
install command also needs to know whether to copy/move the installed
bundle into the profile directory, or just reference it. While we could
use a separate "file operation" `enum` for representing this, it would
have similar downsides to the separate `bool` discussed above: File
operations only make sense for bundle-based sources, so ideally the
types should be designed in a way to not require specifying a file
operation for proxy-based sources. Similarly, the allowed file
operations differ between dev mode and prod mode bundles: Both allow
copy/move, but dev mode also allows references. Therefore, we introduce
the following additional classes to represent sources with file
operations:

- `IwaSourceBundleWithModeAndFileOp`
- `IwaSourceBundleDevModeWithFileOp`
- `IwaSourceBundleProdModeWithFileOp`

Crucially, each of these uses a separate `enum` which only allows
setting the file operations applicable for that class. As file
operations do not apply to `IwaSourceProxy`, we do not introduce
additional classes for it.

- `IwaSourceWithModeAndFileOp` is a thin wrapper around an
  `absl::variant` of `IwaSourceBundleWithModeAndFileOp` and
  `IwaSourceProxy`.
- `IwaSourceDevModeWithFileOp` is a thin wrapper around an
  `absl::variant` of `IwaSourceBundleDevModeWithFileOp` and
  `IwaSourceProxy`.
- `IwaSourceProdModeWithFileOp` is a thin wrapper around an
  `absl::variant` of `IwaSourceBundleProdModeWithFileOp`.
  (`IwaSourceProxy` is always dev mode and thus not part of this
  variant.)

These sources with modes and file ops are used by the newly introduced
`IsolatedWebAppInstallSource` class, which is the new input to the IWA
install command. As part of an upcoming CL, this class will enable us to
set the appropriate `webapps::WebappInstallSource` when installing an
IWA.

I originally prototyped the set of classes representing sources by
making heavy use of inheritance and templating to reduce the amount of
code, but concluded that it was not worth the additional complexity and
downsides (such as it being more verbose to forward-declare a templated
class). I ended up with using a mix of inheritance where possible, and
implicit constructors where inheritance wasn't possible. While the
current code in `isolated_web_app_source.(h|cc)` is a bit verbose and
may seem rather complex at first, I feel like the positive impact it has
on the type-safety of the rest of the code base makes it worth it - its
complexity remains isolated.

  debug code or code that had no existing tests.

Low-Coverage-Reason: LARGE_SCALE_REFACTOR The uncovered code is purely
Fixed: 326044349
Change-Id: Ifd36a8c28ff020ec011f9bd0ef50cd0a5c252b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318031
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272094}