[Sync] Support collaborations in gc_directive
This CL implements garbage collection for collaborations. The server is
expected to provide a list of active collaborations for every
GetUpdates. If some collaboration is not present in the list, all the
previous entities and upcoming updates will be removed / stopped
tracking.

This CL also adds support for collaborations to the fake server but
there are no currently use cases for them. They will be added in
follow-up CLs.

Bug: 325917757
Change-Id: I5a9b73e3e830c0b7500e8003c6d168c701c3cc85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5358039
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272111}