Add sync_pb::PlusAddressSpecifics <-> PlusProfile conversion utils
When receiving a proto from sync, it needs to be converted to a
PlusProfile, to be persisted in the database.

To implement ModelTypeSyncBridge::Get(All)Data(ForDebugging)(), to show
the stored data in sync-internals, we need a way to convert from the
data in the database back to protos.

Bug: b/322147254
Change-Id: I7a4c78611bc34c6ac9d196c5572884d741562426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5355849
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1272122}