Add a use counter for “cursor: hand”.
A ton of web pages use “cursor: hand”, which is a quirks mode hack
for IE5.5 compatibility (it was unneeded from IE6, and removed in IE9),
but nearly all of them have various workarounds so that it doesn't
matter if we remove the functionality. Add a use counter that tries to
capture how often it actually matters.

This is intended to be temporary (perhaps live for one stable release
or similar), so that we get enough data to hopefully remove the
functionality, and then remove the counter again.

Change-Id: Id7c8eab403704fc4bd2ca658a76057497800fb87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366070
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272159}