Kcer: Record PKCS#12 dual-write events
As a part of a migration, PKCS#12 certificates can be dual-written
using Kcer and NSS. Store dual-write events into preferences, so a
rollback can be executed when necessary, if the migration doesn’t
go as planned.

The certificates can be imported from both Ash and Lacros, but the
rollback will always be executed by Ash, so the preference always
needs to be in Ash. Add a method to the CertDatabase mojo interface
to notify Ash when Lacros dual-writes a cert.

Bug: b/264387231
Test: *Kcer*, *CertificateManager* automated tests
Change-Id: I9781b493dd64d6e0a18138c1af7d8c5f0ae506ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5353076
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Michael Ershov <miersh@google.com>
Cr-Commit-Position: refs/heads/main@{#1272167}