Birch: Add suggestion service crosapi for fetching recent tabs
Add a crosapi which can be used to fetch recent tabs in ash when lacros
is enabled. Tabs are initially fetched in SuggestionServiceLacros the
same way we already fetch them in ash, using the OpenTabsUIDelegate.

Add a browsertest to ensure that the crosapi sends data correctly, and
a unittest to ensure that SuggestionServiceLacros fetches tab data.

Bug: b/328640997
Change-Id: I4d0a78aa5483a285d040c01e31d45fb1453052dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5304539
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272182}