[Fuchsia] Create try/fuchsia-x64-cast-receiver-dbg-compile
It's a replica of ci/fuchsia-x64-cast-receiver-dbg but only compile
rather than running the tests. Unlike the -rel variant, it has no
compilator but limits to the changes which are modifying fuchsia
files via location_fitllers to control the overall delay and
resource usage.

Once this try builder is ready to go, the ci|try/fuchsia-x64-dbg can
be removed.

Bug: 1509109
Change-Id: I10b050c15fd84656c6d906faf09ddb527eb18bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362994
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1272186}