Fix auction MojoLPM fuzzer coverage
A promise field was missing `old: 1` -- this caused the auction to wait
forever for that promise, simply "failing" the auction upon destruction
of the TestCase [0].

Also move fields to match the order in the mojom file.

Longer-term, we'll need a way to monitor when we've lost coverage like
this.

[0] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/interest_group/ad_auction_service_impl.cc;l=757;drc=465365641ad68b8feb46a3d23111f221cd2ebc7c

Change-Id: I78f321265b27b751e69950d391e2ac543405ee5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366008
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272209}