Use UserConsentVerifier API for OS authentication on Windows
Changes this CL makes are hidden behind a feature flag.
This CL makes the OS authentication use the new UserConsentVerifier
API. This API works only if user has windows hello configured,
because of that if call to it fails the legacy API will be used.

That change is needed since this new API doesn't block UI thread and
Chrome can still perform its tasks while authentication prompt is
visible. Because of that it aims to reduce lag that happened for some
users after completing authentication using the previous API.

Bug: 327026304
Change-Id: Iec0c5e576bafb784d2aec535a9037ecfc43c3034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5335166
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1272216}