CertProvisioning: Call result callback asynchronously
Call the result callback asynchronously in CertProvisioningWorkers.
The workers might be deleted in the callback while something is
still interacting with them in the current call stack.

Also update the unit tests to wait for the result instead of just
observing the calls, so the tests don’t have to rely on
RunUntilIdle after this CL.

Bug: b:328847969
Test: *CertProv* unit tests, RequiredClientCertificate tast test
Change-Id: I760857e7108a2ac3d1deaebade0a3a48eb81ab3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366208
Commit-Queue: Michael Ershov <miersh@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272229}