[iOS] Populate TabStripMediator consumer with TabGroupItem elements
This CL updates the TabStripMediator so it looks for groups in the
WebStateList, creates a TabGroupItem wrapped in a TabStripItemIdentifier
and adds it before the item associated with the first WebState of that
group. It also observes the WebStateList to repopulate the consumer when
a new group has been added to the WebStateList.

Bug: 328568316
Change-Id: I43e1f2542163d14a553e84e13be98f586c7809da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5350125
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272243}