CrOS Settings: Move ReadWriteCardsManagerImpl initialization
Move the initialization of ReadWriteCardsManager from PostProfileInit to
PreProfileInit to address the dependency issue for the OsSettingsManager
which is created right after user logs and user profile is ready.

SearchSection which is owned by the OsSettingsManager always assumes
that the QuickAnswerController owned by the ReadWriteCardsManager is
always created before itself. Thus moving the initialization of
ReadWriteCardsManager to the PreProfileInit to ensure.

Bug: b/327496668
Change-Id: Ic6975d8a0f983031ab9026da14d22bac2a8355be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5367704
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Connie Xu <conniekxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272286}