[Private Network Access] Get client IP address space from the service worker that loads the main document
This is to fix https://github.com/WICG/private-network-access/issues/126

The issue is that when the main document is fetched by a service worker,
the URLResponseHead's remote_endpoint doesn't have an IP address, which
makes sense because the service worker doesn't always have to forward
the fetch to an actual network fetch.  This results in the render frame
host and the loaders it created having kUnknown IP address space, which
in turn causes the Private Network Access policy to be allow [1].

Also even if the service worker did fetch from file scheme, the service
worker can still modify the content, so the client address space
shouldn't be local, otherwise a public service worker can bypass PNA.

This CL stores the service worker version's IP address space in the
URLResponseHead of the document (main resource) it loads, and then use
that to calculate the document's IP address space.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/private_network_access_util.cc;l=113-117;drc=f4a00cc248dd2dc8ec8759fb51620d47b5114090

Bug: 40241368,323583084
Change-Id: I8751e52cb5e1faa8394531c90d68664e8bd641e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5308035
Auto-Submit: Jonathan Hao <phao@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Jonathan Hao <phao@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272321}