[blink-a11y]: ensure RenderAccessibilityImpl always gets mode changes first
Currently, RenderAccessibilityImpl is an observer of RenderFrameImpl
and reacts to AccessibilitymodeChanged to set up or tear down Blink
accessibility.

This is a problem because, depending on when the observers are added,
RenderAccessibilityImpl may be the *last* observer called. In those
cases, previous observers e.g. PdfAccessibilityTree, may inaccurately
conclude that accessibility isn't on when it *is*, or not have access to
state that should be there e.g. WebAXObjects.

Fix this by making the direct mode changed call from
RenderAccessibilityManager -> RenderAccessibilityImpl (which it owns
anyway).

We also re-enable some PDF a11y tests. These tests were flakey and more
recently, consistently failing after the work in
go/chrome-pdf-a11y-stabilization. It is likely for both cases, this
change found and fixed the root cause of the once flakes now
 fails.

R=aleventhal@chromium.org

Bug: 327458821
Test: cq
Change-Id: I28a63ee6e4149b31f0b0ea097b129eea3a293e1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369374
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1272326}