Update active-map usage in CRD
This CL fixes a few issues we had in our active_map impl.

The first is that we were calling GET_ACTIVEMAP for each frame
but that call is not needed and is likely just there for testing
purposes (the VPX/AOM headers do not describe the usage for it
but I confirmed with the AOM team it's not needed).

The second is pretty minor but instead of creating a new
{aom|vpx}_active_map_t for each frame, I made it a class member.

Lastly, we can skip setting an active map for keyframes but we
still clear it so it's ready for the next frame.

Change-Id: Ia2ae15f63c447d10ae725f8fa6715a9a705228d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366692
Auto-Submit: Joe Downing <joedow@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272339}