Determine ratio of Scroll stages to VSyncs
OnBeginFrame signals from Viz to CC race with the Browser sending Input
to CC. This can lead to inconsistent number of input events per frame
while scrolling. As InputHandlerProxy::DeliverInputForBeginFrame can be
triggered while the input queue is empty. With the input arriving in
InputHandlerProxy::HandleInputEventWithLatencyInfo afterwards.

We want to determine the distribution of when input arrives in the
Renderer vs the VSync times. So we can experiment with waiting for
input when the queue is empty. We also want to determine the
distribution of event generation times to judge thresholds to not
pause waiting for.

Bug: 40132248
Change-Id: Ia151ed3390fa9226923019aebfacb46f6c18ace2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5367303
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Victor Miura <vmiura@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272342}