Move static part of upload preparation.
Part of the upload requests that never change are moved to the client,
so that they don't need to be kept together with each cached upload.
With that, we can cache the events and perform upload to the server
at a later time, if necessary.

Tested manually on VM and with
`git cl try -B luci.chrome.try -b win-celab-try-rel`

Bug: b:327243582
Change-Id: I1b9a9d567c6e5585e042e9de8194728c5aa845ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362604
Reviewed-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272368}