Update layer tree of windows in overview-mode to avoid render surfaces.
To avoid the need for a render surface to draw to overview window
correctly, we need to update the rounded corners of the layer tree of
the overview window, as discussed in crrev.com/c/5296786.

This CL also add the functionality to restore the rounded corners
of the layer tree.

Bug: b/326468096, b/325635179
Change-Id: Id6466db461ad4abc42b781a37b3829f274bb5c49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5334038
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272410}